requirements.txt
fastapi
uvicorn
sendgrid
requests
python-dotenv
pytz
